FEATURE(api): APIs are divided into modules
