+++ app/blueprints/api/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+api = Blueprint('api', __name__, url_prefix='/api')
+
+from app.blueprints.api import auth, task, admin
